rename mit ocw data science folder names; update path in tests lec2
