+++ src/mit_ocw_data_science/lec2/menu.py
