+++ src/mit_ocw_data_science/lec3/60002-lecture3.py
