+++ src/mit_ocw_data_science/lec3/README.md
