+++ src/mit_ocw_data_science/lec5/plot_matplotlib.py
