+++ tests/team_ta/test_food.py
@@ -1,6 +1,6 @@
 import pytest
-
-from menu import Food
+import add_path
+from mit_ocw_data_science.lec2.menu import Food
 
 @pytest.fixture
 def food():
