+++ tests/team_ta/test_greedy.py
@@ -1,6 +1,7 @@
 import random
 import pytest
-from menu import *
+import add_path
+from mit_ocw_data_science.lec2.menu import *
 
 def test_greedy():
     items = [
